Incorporate ball-ball friction model
