+++ pooltool/physics/resolve/ball_ball/frictional_mathavan/__init__.py
@@ -9,6 +9,10 @@ from numpy.typing import NDArray
 import pooltool.constants as const
 from pooltool.objects.ball.datatypes import Ball, BallState
 from pooltool.physics.resolve.ball_ball.core import CoreBallBallCollision
+from pooltool.physics.resolve.ball_ball.friction import (
+    AlciatoreBallBallFriction,
+    BallBallFrictionStrategy,
+)
 from pooltool.physics.resolve.models import BallBallModel
 
 INF = float("inf")
@@ -233,6 +237,7 @@ class FrictionalMathavan(CoreBallBallCollision):
     """
 
     model: BallBallModel = attrs.field(default=BallBallModel.FRICTIONAL_MATHAVAN)
+    friction: BallBallFrictionStrategy = AlciatoreBallBallFriction()
     num_iterations: int = 1000
 
     def solve(self, ball1: Ball, ball2: Ball) -> Tuple[Ball, Ball]:
@@ -260,7 +265,7 @@ class FrictionalMathavan(CoreBallBallCollision):
             u_s1=ball1.params.u_s,
             u_s2=ball2.params.u_s,
             # Assume the interaction coefficients are the average of the two balls
-            u_b=(ball1.params.u_b + ball2.params.u_b) / 2,
+            u_b=self.friction.calculate_friction(ball1, ball2),
             e_b=(ball1.params.e_b + ball2.params.e_b) / 2,
             N=self.num_iterations,
         )
